JarakSolusiIdeal

+++ database/migrations/2022_10_13_071310_create_matrices_table.php
@@ -16,11 +16,11 @@ return new class extends Migration
         Schema::create('matrices', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('alternative');
-            $table->integer('c1');
-            $table->integer('c2');
-            $table->integer('c3');
-            $table->integer('c4');
-            $table->integer('c5');
+            $table->float('c1');
+            $table->float('c2');
+            $table->float('c3');
+            $table->float('c4');
+            $table->float('c5');
             $table->timestamps();
         });
     }
